more renaming and deprecation warnings added
