+++ src/main/app/src/app/formulieren/taken/model/default-taakformulier.ts
@@ -10,6 +10,7 @@ import { ReadonlyFormFieldBuilder } from "../../../shared/material-form-builder/
 import { TextareaFormFieldBuilder } from "../../../shared/material-form-builder/form-components/textarea/textarea-form-field-builder";
 import { AbstractTaakFormulier } from "../abstract-taak-formulier";
 
+/** @deprecated Migrate to Angular task form pattern (AbstractTaskForm) */
 export class DefaultTaakformulier extends AbstractTaakFormulier {
   fields = {
     REDEN_START: "redenStart",
